Decouple DirectActionInterface from PdoAdapter
